+++ packages/react/src/lib/components/Plots/Pie/StackedDonut/StackedDonutBase.tsx
@@ -14,8 +14,8 @@ import { useSelector } from "react-redux";
 import { useLegendItems, useRender } from "../../../../hooks";
 
 import { renderCanvas } from "../../renderCanvas";
-import type { IArcAngles } from "../interpolateArc";
-import { interpolateArc } from "../interpolateArc";
+import type { IArcAngles } from "../../interpolateArc";
+import { interpolateArc } from "../../interpolateArc";
 import { useFocused } from "../../useFocused";
 import { useTooltip } from "../../useTooltip";
 import { useZoom } from "../../useZoom";